JsonString[Indent]: Remove trailing \n
By using the custom encoder to get rid of the html encoding, we need to use enc.Encode which adds a \n to the output. That is different from what json.Marshal[Indent] produces and to create the same output as before (apart from the HTML encoding), we remove the \n after encoding.